apps/fast-api/pyproject.toml
[project]
name = "skillhub-api"
version = "1.0.0"
description = "SkillHub FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]",
  "pydantic-settings",
  "sqlalchemy>=2.0",
  "pyjwt",
  "psycopg2-binary",
  "alembic",
  "httpx",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-exporter-otlp-proto-grpc",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-instrumentation-sqlalchemy",
  "opentelemetry-instrumentation-httpx",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "pytest-asyncio",
  "httpx",
  "mypy",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.11"
